schema: Sort results by partition keys in multi-partition tests
The order of "IN" queries for multiple partitions is different in
Cassandra and Scylla:

https://github.com/scylladb/scylla/issues/2029

Gemini finds this problem, of course, but since this is a known issue,
let's work around it by sorting the results explicitly (suggested by
Avi).

We can revert the change when the problems is fixed in upstream Scylla.