schema: Restrict partition key type to int
We partition keyspace between different concurrent goroutines to ensure
there are no concurrent writes when we read. However, as pointed out by
Henrik, the partition keys are mapped to tokens and there are no
guarantees that the mapping respects the partitioning for arbitrary
types, which can lead to a concurrent write on the partition we are
reading from.

Threfore, restrict partition key type to int, which fixes
the Gemini complaint that oracle and test cluster results sets have
different sizes.