release: injected vars in place
The version info can be printed using the '--version' program argument.
The current solution works out of the box with 'goreleaser' which
will find the latest tag if there is such and the commit that it
refers to. The build date is also injected and printed along with
the commit information.