schema: collection types added
This brings in sets, lists, maps as well as their frozen counterparts.
Provisions had to be made for indexing support and a method Indexable
is added to the Type interface.

The types are all moved to a new file to avoid further cluttering of the
schema file.

Diffing needed extension to make maps flattened and sorted before
comparison.