gemini: separate store interface
The session is removed and a store interface is introduced.
This store interface is small and isolated and allows for the
caller to generate arbitrary queries using the builder.

The store encapsulates the two internal stores and performs
checks using data from these two instances.

Currently the internal stores are both cql based but any
backing storage can be implemented without affecting the
surrounding calling code.