gemini: terminating properly upon an error
We ensure that we drain the remaining work after a quit signal is
detected. This ensures that no workers get blocked on trying to
send a status message after the receiving end has stopped listening.

We also send down the termination context to the driver to allow
it to abort what it is doing as fast as possible.