gemini: prometheus metrics added
Gemini now exposes prometheus metrics via http on the path '/metrics'.
The gemini specific metrics is so far CQL operations labeled as either
'batch', 'delete', 'insert', 'select' or 'update'.

The general Go runtime prometheus properties for the runtime are
also exposed which can be useful for seeing how for example the GC
is behaving.