gemini: warmup time configurable using cli argument
Gemini now has a cli argument 'warmup' which is a duration given
in a format such as '30s', '15m' or '10h'.
This time is added to the duration time so as to make total run
time easy to calculate.

For example: Given a duration of 10h and a warmup of 2h the total
runtime will be an 12 hours and validations will start to happen
after a 2 hours warmup run with only inserts.