gemini: user can select compaction strategy
A CLI argument "compation-strategy" is added which can be supplied
in a number of forms.

Either as a set of string values "stcs", "twcs" or "lcs" which will
make Gemini choose the default values for the properties of the
respective compation strategies.

The second option is to simply add the entire JSON-like definition
"{"class"="SizeTieredCompactionStrategy", "enabled"=true, ....}".
This will be then be applied to the created tables.

NB:
The format of this JSON-like data needs to be actual valid JSON.
This means it needs to have double quotes instead of single and
numbers and booleans should not have any quptes at all.