schema: ddl jobs added
Mutation jobs now also emit DDL jobs to allow testing with schema
changes. Currently only colum add and drop are supported due to
https://issues.apache.org/jira/browse/CASSANDRA-12443 and the
frequency of the schema changes are not configurable.

The synchronization of the internal Table struct is handled with
explicit locking which may not be ideal but a message passing
solution would require extensive refactoring.