gemini: allows inflight ops to finish
A "pump" concept is introduced with the purpose of being the force
driving the operations against the stores. This pump exposes a
channel that will continously emit heartbeats that has an optional
sleep parameter telling the currently operating goroutine to sleep
for the provided duration. When the pumps channel is closed all
listening jobs will finish the currently available heartbeats in a
race free manner. To speed up the process, the pump also tries to
drain any remaining heartbeats.