schema: data sizes are configurable
The data sizes are configurable through the SchemaConfig object.
It propagates most of config through to the runtime at the initial
schema generation. However the sizes of actual data such as strings
and blobs are carried through to the GenValue functions via the
PartitionRange objects.

Default data set size is "large" which is suitable for a large actual
testing setup and the development scripts is using the new switch
"dataset-size" to set it to "small" during development.