store: oracle mutations more resilient
Gemini used to treat mutation errors for both systems as equal.
Now gemini treats failures in the Oracle as unrecoverable and it
will simply not attempt the same mutation in the test system.

The idea is that since the Oracle is the blueprint, there is no
need to to attempt a test mutation if the blueprint didn't change.