gemini: reattempt partially empty loads of secondary index results
Since secondary index building happens in the background they are
not assured to be available directly after the insert happens.
We try to mitigate this by simply running the query again for
a few times until a max attempts limit is reached or a non-zero
result set is achieved for both of the sets.