store: don't enable tracing if not requested
The `newSession` function compared `out` (the tracing output
destination) with `nil` to decide if tracing should be enabled on the
created session.

Although the underlying value of `out` was `nil *os.File`, it was passed
in as `io.Writer`. However, `nil *os.File` is a non-`nil` `io.Writer`,
so the check always passed.

Fixes #247.