fix(utils): generate time values in the correct range
According to the CQL binary protocol, time is an int64 in range [0;86399999999999]

https://github.com/apache/cassandra/blob/f5df4b219e063cb24b9cc0c22b6e614506b8d903/doc/native_protocol_v4.spec#L941
An 8 byte two's complement long representing nanoseconds since midnight.
Valid values are in the range 0 to 86399999999999

Values outside of this range are invalid.

RandTime() generated values using rnd.Int63(),
but this generates values outside of the allowed range.

Let's fix it so that it generates values in the correct range.

Fixes: https://github.com/scylladb/scylladb/issues/14645
Refs: https://github.com/scylladb/scylladb/issues/14667

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>