feature(materialized-views): add cli flags to disable materialized views
There is a crash in gemini, which is deeped then we expected first, and
it comes down to the materialized views generation and validation.
Something there is wrong, until further investigation (and possibly
gemini redisign, this issue cannot be fixed). As I've discussed with
@fruch about this issue, we concluded that the best way forward for now
is to disable materialized views until the redisign of the project.

Simplest solution is just to add the CLI flag and propagate it to the
creation of the materialized views.

Signed-off-by: Dusan Malusev <dusan.malusev@scylladb.com>