fix(prettyCQL): prettyCQL crashes on ouf-of-bound access
When gemini tries to `pretty print` CQL, in some cases can fail
with `out-of-bounds` access on array.

```go
out = append(out, queryChunks[qID])
```

This fix uses new go1.23 iter funcs, which removes the allocations,
and is not concerened with indexes which caused the issue in the first
place.

I tried to remove the `switch .(type)` but `TuppleType` is different
so has to be still there.

Signed-off-by: Dusan Malusev <dusan.malusev@scylladb.com>