improvement(stmt): refactoring statement genration
Statement generation currently does not work as intended,
gemini generates values for primary and clustering keys,
that are completply invalid. The are not just off by value,
but also off by type, e.g. generates `text` for `timeuuid`.
This is the first step in resolving this issue, by refactoring
how, when and where statements are generated.

Signed-off-by: Dusan Malusev <dusan.malusev@scylladb.com>