refactor(statement-logger): statement logger data-racy
Statement Logger had a flow, it assumed it has the owner ship of the
query, but statements are reused this means, it produced queries that
are not even possible to execute. (This does not fix the issue of more
columns expected then returned). Since logger works as single thread
to write to the file (Queue), it has to be reworked to format the query
and the values before they are passed to the channel (works like a
queue), this way a query is owned by the channel inside a bytes.Buffer.

PrettyCQL has been rewritten to use `bytes.Buffer` instead of
`strings.Builder`, as the intent of the PrettyCQL is to be stored inside
a file, and files work with `[]byte` rather then strings.

Signed-off-by: Dusan Malusev <dusan.malusev@scylladb.com>