refactor(filelogger): make filelogger return a error instead of void
Up until this point, we have been ignoring the errors that the filelogger, from now on, we will return the error and let the caller handle it. This makes the statement that needs to be executed, run first (both at oracle and at test cluster), after that we log the statement that was executed.
If the statement cannot be logged, operation will be aborted and the error will be returned.

Signed-off-by: Dusan Malusev <dusan.malusev@scylladb.com>