refactor(stopFlag): completly remove stopFlag for `context`
`context` is Go package used mainly to signal cancelation.
`stopFlag` package inside of gemini was used the same way.
It was just adding overhead and maintainability issues in the long
run, as it was only a wrapper for `context.Context` and all at once
calling the `cancel` funcs at the end of the program. This removal
makes code easier to follow for everybody working on it. `context`
package is well known in Go community and can serve many purposes
beyond just canceling the background workers (`jobs` in gemini
terminology).

Hard kill in gemini was used to signal the immadiet stoppage
for gemini, without pritting the results, and soft on the contrary
could be triggers in a few ways:

- gemini validation failed
- gemini mutation failed
- SIG(INT,TERM) was sent

There is no reason to have HARD kill in the application, if we need hard
kill, SIGKILL can be sent and everything would be stopped, there will be
no need to the the cleanup as kernel will ensure that happens.

`context.Context` works as a soft kill, and if something happens bad
in gemini (validation fails or mutation fail) `globalStatus.HasError()`
would stop all other `goroutines` from continuing if `failFast` CLI flag
is passed, so the set of `softKill` is not needed.

Signed-off-by: Dusan Malusev <dusan.malusev@scylladb.com>