feature(random): use math/rand/v2 instead of exp/rand
golang.org/x/exp/rand package is depricated in 1.24.
It has been replaces with better math/rand/v2, the first
v2 in the STD library.

math/rand/v2 is better in every scenario and more performant.
Default source is PCG, which is, by all the metrics, better by
all metrics in terms of randomness and speed.

Signed-off-by: Dusan Malusev <dusan.malusev@scylladb.com>