fix(patitions): too many stale partitions
On some seeds the keypspace generated has a partition key which has
number of values too low and they generate the same token range puting
them only in few buckets. This will make gemini too many retries to get
the value from the partition if the distribution configured is not
evenly spread. Like in the case of seed 60, from 10k partitions, number
of state partitions will be >95% due to the partition key having low
number of values (as explained). This effect causes gemini to stall and
practically never finish, since the retries will take longer and when
exiting, channels will close and reading from closed channel is blocking
action in the current codebase. To avoid this bug, the number of
partitions keys has been raised to 2 instead of 1, making it highly
unlikely to have this bug in production. This is just a temporary fix
until we decide what we need to do with gemini partitions.

Signed-off-by: Dusan Malusev <dusan@dusanmalusev.dev>