fix(stmtlogger): missing statements in log files in both oracle and test (#562)
Since the Mutation Fragments and CQL Statements cannot share the same
code for fetching them (it was before), this made issue that statements
are ignored and always empty in the log file. Since this is a core
feature and important for us to tell what happened if gemini failed, the
initialization of Connection to both Oracle and Test cluster had to be
moved later (after the statement logger is setup) due to how statement
logger works.

This fixes the issue, and comprehensive test suite is added to ensure
this feature will not be broken in the future. These tests cover,
insert, select statements and select mutation fragments from scylla.

Signed-off-by: Dusan Malusev <dusan@dusanmalusev.dev>