fix(stmtlogger): small channel, low worker count, and network overhead stalled throughput
Small buffer size in the channel for statement logger

The statement logger has channel which sends data to many threads for insertion inside the oracle's statement_logger table, when this channel gets full, it blocks and the gocql aparatus cannot move forwards (cause gocql mechanisms of query obervations are used to log the data that is mutation). This makes the connection blocked until the new item can be inserted. This explains why at some point in time gemini stabilizes the throughput around ~2k req/s.

Not enough workers for the statement logger

While inserting into the statement logger's table, the number of workers were too low to accomodate the number of mutation threads that were doing the queries. The new solution implemented will autoscale the number of workers when the channel gets filled and stays filled. This is capped at 128 workers so that it does not take too much CPUs (as it is IO bound, higher amount of workers is ok) from the mutation and validations.

Network bottleneck

Since inserting into the oracle takes not just time, but networking, and oracle is getting 3x more inserts then test, this becomes an issue. (3x comes from - 1 insert that is the same with test cluster, 1 insert into statement logger for test mutation and 1 insert into statement logger for oracle mutation). To make it better, and only 1 round trip, batching of the insertions into the statement logger table is neccessary. Small buffer (128 items size) keeps the items in memory and after it's filled it is sent to Scylla as a Batch Request.