feature(partitions): lock per partitions instead  of whole partitions
Performance degrades when selecting a partition for READ and for WRITE,
since it had a global locking mechanism on all of the partitions. This
makes lock contention really high, since DELETE statement need to remove
the particular partition and replace it with new one for further use
(number of partitions in memory always remains constant). If we have 500
delete req/s, that's 500 write locks on whole array, meaning no other
operation can progress.

To mitigate this issue, each partition now contains it's own lock for
reading and writing - still only delete takes the WRITE lock on the
partition to replace it but not on the whole array of partitions, only
on selected one, making possible other validation threads to move
forward on other partitions since they are not locked. In some rare
cases, this issue can still happen (e.g small amount of partitions <
1000) where lock contention for one partition is high, but in general
      case this lock-per-partition solves the problem